feat: created sub-commands for download and transform and fixed a schema bug
