+++ src/bundestag/cli/utils.py
@@ -1,0 +1,3 @@
+import typer
+
+OPTION_DRY = typer.Option(False, help="Dry or not")
